refactor: split cli_dump_helpers.py's depth resolvers into cli_dump_depth.py
Merging PR 3A's two migrations (legacy compile-DB match unfolding, scan
candidate resolution routing) pushed cli_dump_helpers.py to 2012 lines,
over the AI-readiness 2000-line hard cap. Moved resolve_dump_depth/
resolve_dump_collect_context to a new leaf module, cli_dump_depth.py --
pure behavior-preserving split, both names still resolve from
abicheck.cli_dump_helpers via an explicit re-export.
